// src/components/MapPartComponents/data.d.ts
export interface StateType {
    dmCodeList?:Array<object>,
    gridPointList?:Array<object>,
    portTypeList?:Array<object>,
    ddlList?:Array<object>,
    pathList?:Array<object>,
    nowAgvState?: Array<object>,
    socketObj?:Object,
    exceptionList?: Array<object>,
    socketStatus?: String,
}